computer_club_server/requirements.txt
# Веб-фреймворк
fastapi==0.104.1
uvicorn[standard]==0.24.0

# База данных
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0  # для PostgreSQL

# Валидация данных
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Аутентификация
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Кэширование и очереди
redis==5.0.1
celery==5.3.4

# HTTP клиент
httpx==0.25.1

# Переменные окружения
python-dotenv==1.0.0